Move `hashGraph` -> `graph`
