+++ packages/hash/integration/src/tests/model/ontology/propertyType.model.test.ts
@@ -1,5 +1,5 @@
 import { getRequiredEnv } from "@hashintel/hash-backend-utils/environment";
-import { createGraphClient } from "@hashintel/hash-api/src/hashGraph";
+import { createGraphClient } from "@hashintel/hash-api/src/graph";
 import { Logger } from "@hashintel/hash-backend-utils/logger";
 
 import { PropertyType as PropertyTypeSchema } from "@hashintel/hash-graph-client/";
